Some simple pathfdinding
The enemy doesn't actually use this pathfinding because it doesn't work
yet. I'm pushing just to show progress and also so that other people can
check my logic if they want to do so.

My next goal is to make it so that if an enemy can see an exit they will
go to it at all cost.

Pathing away from snake not implemented